server-flask/requirements.txt
flask
flask-cors
flask-restful
flask-mail
pymongo[srv]
certifi
dnspython
python-dotenv
pyjwt
bcrypt
cloudinary
reportlab
pandas
gunicorn
qrcode[pil]
pillow
